server/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.4.1
asttokens==2.2.1
async-timeout==4.0.3
backcall==0.2.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.0
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==35.0.0
decorator==5.1.1
distlib==0.3.6
Django==3.2.8
django-cors-headers==3.10.0
django-csp==3.7
django-filter==23.2
django-redis==5.3.0
django-webpack-loader==2.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
executing==1.2.0
filelock==3.12.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
identify==2.5.23
idna==3.4
ipython==8.14.0
jedi==0.19.0
kombu==5.3.2
Markdown==3.4.3
matplotlib-inline==0.1.6
nodeenv==1.7.0
numpy==1.25.2
oauthlib==3.2.2
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==3.5.0
pre-commit==3.2.2
prompt-toolkit==3.0.39
protobuf==4.24.1
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.20
Pygments==2.16.1
PyJWT==2.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.1
pytz==2021.3
PyYAML==5.3.1
redis==4.6.0
reportlab==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==3.10.0.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6